\name{ICoVeR-package}
\alias{ICoVeR-package}
\alias{ICoVeR}
\docType{package}
\title{Interactive Contig-bin Verification and Refinement}
\description{
Provides Parallel Coordinates for R Data frames.
}
\details{
\tabular{ll}{
Package: \tab ICoVeR\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-11-04\cr
License: \tab GPLv3\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
  Bertjan Broeksema <research@broeksemaatjes.nl>,
  Magdalena Calusinska <magdalena.calusinska@list.lu>,
  Fintan McGee <fintan.mcgee@list.lu>,
  Klaas Winter,

  Maintainer: Bertjan Broeksema <research@broeksemaatjes.nl>
}
\references{}

\keyword{~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~}

\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
